Intake Program Was Made
We Made a Intake program that uses the right joystick x value on gamepad1 to control the power of the motor spinning., and we made it so that the power can't be less than 0 and greater than 1.  We also used telemetry to see what value of power the robot is getting. The reason we made this is to test a bunch of different intake designs to see  which one outperforms all the others.